Move constants to constants folder
