+++ src/constants/Constants.ts
@@ -11,3 +11,7 @@ export default function ITREXVARS(): ITREX_VARIABLES {
     const releaseChannel = Constants.manifest.extra.variables;
     return releaseChannel;
 }
+
+export class ITRexVariables {
+    RELEASE_CHANNEL = Constants.manifest.extra.variables;
+}
